third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/hidden-elements.html

<!doctype html>
<link rel=help href="https://html.spec.whatwg.org/#hidden-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div hidden></div>
<script>
const kLocalNames = [
  "area", "base", "basefont", "datalist", "head", "link", "meta", "noembed",
  "noframes", "param", "rp", "script", "source", "style", "template",
  "track", "title",
];

for (let name of kLocalNames) {
  test(function() {
    let element = document.createElement(name);
    document.body.appendChild(element);
    assert_equals(getComputedStyle(element).display, "none");
  }, `${name} should be hidden`);
}

test(function() {
  assert_equals(getComputedStyle(document.querySelector("[hidden]")).display, "none");
}, `[hidden] element should be hidden`);
</script>
