third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/hidden-elements-expected.txt

This is a testharness.js-based test.
FAIL area should be hidden assert_equals: expected "none" but got "inline"
FAIL base should be hidden assert_equals: expected "none" but got "inline"
FAIL basefont should be hidden assert_equals: expected "none" but got "inline"
PASS datalist should be hidden
PASS head should be hidden
PASS link should be hidden
PASS meta should be hidden
FAIL noembed should be hidden assert_equals: expected "none" but got "inline"
PASS noframes should be hidden
PASS param should be hidden
PASS rp should be hidden
PASS script should be hidden
FAIL source should be hidden assert_equals: expected "none" but got "inline"
PASS style should be hidden
PASS template should be hidden
FAIL track should be hidden assert_equals: expected "none" but got "inline"
PASS title should be hidden
PASS [hidden] element should be hidden
Harness: the test ran to completion.
